requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
psycopg2-binary
alembic
pydantic>=2.0
python-jose[cryptography]
passlib[bcrypt]
jinja2